web/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.28.0
Flask==1.1.2
Flask-SocketIO==4.3.1
greenlet==0.4.17
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
pika==1.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
